renamed: src/components/aside/index.js -> src/components/aside/desktop-aside.js
